+++ src/model/dividends.model.js
@@ -1,6 +1,9 @@
 import knex from "../db";
 import dividendsType from "../enum/dividendsType";
+import { jsonObjectQuerySelect } from "../utils";
 import transacting from "../utils/transacting";
+import * as investmentModel from "./investment.model";
+
 const TABLE_NAME = "dividends";
 export const selectDefault = [
     "id",
@@ -42,23 +45,7 @@ export const selectDefault = [
 export const findAll = (options, trx) => {
     const query = knex(TABLE_NAME)
         .select([
-            knex.raw(`
-                JSON_OBJECT(
-                    'id', investment.id,
-                    'name', investment.name, 
-                    'longName', investment.longName, 
-                    'balance', investment.balance, 
-                    'sector', investment.sector, 
-                    'previousClosePrice', investment.previousClosePrice, 
-                    'volumeDay', investment.volumeDay, 
-                    'changePercentDay', investment.changePercentDay, 
-                    'priceDay', investment.priceDay, 
-                    'priceDayHigh', investment.priceDayHigh, 
-                    'priceDayLow', investment.priceDayLow, 
-                    'createdAt', investment.createdAt, 
-                    'updatedAt', investment.updatedAt
-                ) as investment
-            `),
+            knex.raw(jsonObjectQuerySelect("investment", investmentModel.selectDefault)),
             ...selectDefault.map((select) => {
                 return `${TABLE_NAME}.${select}`;
             })
@@ -94,23 +81,7 @@ export const findAll = (options, trx) => {
 export const findUpdateDivideds = (date, trx) => {
     const query = knex(TABLE_NAME)
         .select([
-            knex.raw(`
-                JSON_OBJECT(
-                    'id', investment.id,
-                    'name', investment.name, 
-                    'longName', investment.longName, 
-                    'balance', investment.balance, 
-                    'sector', investment.sector, 
-                    'previousClosePrice', investment.previousClosePrice, 
-                    'volumeDay', investment.volumeDay, 
-                    'changePercentDay', investment.changePercentDay, 
-                    'priceDay', investment.priceDay, 
-                    'priceDayHigh', investment.priceDayHigh, 
-                    'priceDayLow', investment.priceDayLow, 
-                    'createdAt', investment.createdAt, 
-                    'updatedAt', investment.updatedAt
-                ) as investment
-            `),
+            knex.raw(jsonObjectQuerySelect("investment", investmentModel.selectDefault)),
             ...selectDefault.map((select) => {
                 return `${TABLE_NAME}.${select}`;
             })