+++ src/utils/index.js
@@ -62,3 +62,19 @@ export const parseStringToDividendType = (value)=>{
             return dividendsType.YIELD;
     }
 };
+
+/**
+ * 
+ * @param {string} join 
+ * @param {Array<string>} selects
+ * @returns {string} 
+ */
+export const jsonObjectQuerySelect = (join, selects)=>{
+    const data = [];
+
+    selects.forEach((select)=>{
+        data.push(`'${select}', ${join}.${select}`);
+    });
+
+    return "JSON_OBJECT("+ data + `) as ${join}`;
+};